+++ packages/base/src/base-element.ts
@@ -27,6 +27,7 @@ export {
 } from '@material/base/types.js';
 export {addHasRemoveClass} from './utils.js';
 
+/** @soyCompatible */
 export abstract class BaseElement extends LitElement {
   /**
    * Root element for MDC Foundation usage.