+++ packages/button/src/mwc-button.ts
@@ -20,6 +20,7 @@ import {ButtonBase} from './mwc-button-base.js';
 import {style} from './mwc-button-css.js';
 import {withStyles} from '@material/mwc-base/base-element.js';
 
+/** @soyCompatible */
 @customElement('mwc-button')
 export class Button extends ButtonBase {
   static styles = style;