+++ packages/checkbox/src/mwc-checkbox.ts
@@ -26,6 +26,7 @@ declare global {
   }
 }
 
+/** @soyCompatible */
 @customElement('mwc-checkbox')
 export class Checkbox extends CheckboxBase {
   static styles = style;