+++ packages/formfield/src/mwc-formfield-base.ts
@@ -1,30 +1,35 @@
 /**
-@license
-Copyright 2018 Google Inc. All Rights Reserved.
+ * @license
+ * Copyright 2018 Google Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+// tslint:disable:no-new-decorators
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-import {MDCFormFieldAdapter} from '@material/form-field/adapter.js';
-import MDCFormFieldFoundation from '@material/form-field/foundation.js';
-import {BaseElement, EventType, SpecificEventListener} from '@material/mwc-base/base-element.js';
-import {FormElement} from '@material/mwc-base/form-element.js';
-import {observer} from '@material/mwc-base/observer.js';
-import {findAssignedElement} from '@material/mwc-base/utils.js';
+import {MDCFormFieldAdapter} from '@material/form-field/adapter';
+import MDCFormFieldFoundation from '@material/form-field/foundation';
+import {BaseElement, EventType, SpecificEventListener} from '@material/mwc-base/base-element';
+import {FormElement} from '@material/mwc-base/form-element';
+import {observer} from '@material/mwc-base/observer';
+import {findAssignedElement} from '@material/mwc-base/utils';
 import {html, property, query} from 'lit-element';
-import {classMap} from 'lit-html/directives/class-map.js';
+import {classMap} from 'lit-html/directives/class-map';
+
 
 export class FormfieldBase extends BaseElement {
   @property({type: Boolean}) alignEnd = false;
+  @property({type: Boolean}) spaceBetween = false;
+  @property({type: Boolean}) nowrap = false;
 
   @property({type: String})
   @observer(async function(this: FormfieldBase, label: string) {
@@ -61,7 +66,7 @@ export class FormfieldBase extends BaseElement {
         if (input instanceof FormElement) {
           const ripple = await input.ripple;
           if (ripple) {
-            ripple.activate();
+            ripple.startPress();
           }
         }
       },
@@ -70,7 +75,7 @@ export class FormfieldBase extends BaseElement {
         if (input instanceof FormElement) {
           const ripple = await input.ripple;
           if (ripple) {
-            ripple.deactivate();
+            ripple.endPress();
           }
         }
       },
@@ -90,10 +95,14 @@ export class FormfieldBase extends BaseElement {
   }
 
   protected render() {
+    const classes = {
+      'mdc-form-field--align-end': this.alignEnd,
+      'mdc-form-field--space-between': this.spaceBetween,
+      'mdc-form-field--nowrap': this.nowrap
+    };
+
     return html`
-      <div class="mdc-form-field ${classMap({
-      'mdc-form-field--align-end': this.alignEnd
-    })}">
+      <div class="mdc-form-field ${classMap(classes)}">
         <slot></slot>
         <label class="mdc-label"
                @click="${this._labelClick}">${this.label}</label>