+++ packages/icon-button/src/mwc-icon-button.ts
@@ -27,6 +27,7 @@ declare global {
   }
 }
 
+/** @soyCompatible */
 @customElement('mwc-icon-button')
 export class IconButton extends IconButtonBase {
   static styles = style;