+++ packages/list/src/mwc-radio-list-item-base.ts
@@ -46,8 +46,10 @@ export class RadioListItemBase extends ListItemBase {
         this.renderGraphic() :
         html``;
     const meta = this.hasMeta && this.left ? this.renderMeta() : html``;
+    const ripple = this.renderRipple();
 
     return html`
+      ${ripple}
       ${graphic}
       ${this.left ? '' : text}
       <mwc-radio