+++ test/src/util/helpers.ts
@@ -17,6 +17,10 @@ limitations under the License.
 import {customElement, html, LitElement, property} from 'lit-element';
 import {render, TemplateResult} from 'lit-html';
 
+interface HasKeyCode {
+  keyCode: number;
+}
+
 declare global {
   interface Window {
     tachometerResult: undefined|number;
@@ -80,12 +84,14 @@ export class TestFixture extends LitElement {
 
 const defaultOpts = {
   shouldAttachContents: true,
-  document: document
+  document: document,
+  afterRender: null,
 };
 
 interface FixtureOptions {
   shouldAttachContents: boolean;
   document: Document;
+  afterRender: ((root: ShadowRoot) => Promise<void>)|null;
 }
 
 export const fixture =
@@ -99,6 +105,10 @@ export const fixture =
   opts.document.body.appendChild(tf);
   if (opts.shouldAttachContents) {
     await tf.updateComplete;
+  }
+
+  if (opts.afterRender) {
+    await opts.afterRender(tf.root);
   }
 
   return tf;
@@ -167,3 +177,39 @@ export const measureFixtureCreation = async (
 export const rafPromise = async () => new Promise((res) => {
   requestAnimationFrame(res);
 });
+
+export class Fake<TArgs extends any[], TReturn> {
+  public calls: Array<{args: TArgs}> = [];
+  public get called(): boolean {
+    return this.calls.length > 0;
+  }
+  public get callCount(): number {
+    return this.calls.length;
+  }
+  public returnValue?: TReturn;
+  public handler: (...args: TArgs) => TReturn;
+
+  public constructor() {
+    this.handler = (...args: TArgs) => {
+      this.calls.push({args});
+      return this.returnValue as TReturn;
+    };
+  }
+}
+
+export const waitForEvent = (el: Element, ev: string) => new Promise((res) => {
+  el.addEventListener(ev, () => {
+    res();
+  }, {once: true});
+});
+
+export const ieSafeKeyboardEvent = (type: string, keycode: number) => {
+  // IE es5 fix
+  const init = {detail: 0, bubbles: true, cancelable: true, composed: true};
+  const ev = new CustomEvent(type, init);
+
+  // esc key
+  (ev as unknown as HasKeyCode).keyCode = keycode;
+
+  return ev;
+}